fixes to SRAM and a dumb crash
I was testing on BlastEm, and I assumed it didn't like the SRAM implementation. Turns out the bug that started this was a stupid assembler bug in LoadLifeIcon, but the SRAM had a crash of its own along with some obvious bugs. I've fixed all of that.
Game should work good now, though you might need to delete your current saves to work the new code. Public releases should probably have a method for converting the older SRAM types if we do any reorganizing, but for now it's not a big deal
Still don't know what's going on with the menu on Fusion though